se elimino la tecnologia expo y reacct native, se agrego un formulario para cargar los dispensers en el mapa

+++ backend/src/apps/core/models.py
@@ -8,5 +8,5 @@ class Imagen(models.Model):
 
 class Ubicacion(models.Model):
     codigo_ubicacion = models.BigAutoField(primary_key=True)
-    longitud = models.IntegerField()
-    latitud = models.IntegerField()
+    longitud = models.FloatField()
+    latitud = models.FloatField()
